Moved exploration utilities to a library
